Initial commit - Added index.html for portfolio
This is my first version of the personal portfolio website built using HTML, CSS, and JavaScript. It includes a responsive layout and is designed to showcase my skills as a web developer and designer. The website serves as an introduction to my work and provides contact information for freelance opportunities.